Added Freetype support for Windows platforms
This commit comes from a patch submitted to the qt dev team:

  https://codereview.qt-project.org/100430

The patch was reviewed and rejected because Qt4 was in maintenance mode and this was quite a large change.
Nevertheless, the patch has been used successfully with both MSVC 2008 and MSVC 2010.
In particular, it has been used to create Qt opensource builds for Windows CE using VS2008.

The patch was based on changes in Qt5 to support Freetype on Windows.

================

Known problems / concerns:

- Some Windows code under examples does not build due to assumptions
- about the FontEngine.  Just needs fixing..
- text.pri is a mess
- Had to modify one line in Freetype source bundled with Qt.
- Did I add leaky abstractions???
- MS's 2008 CRTL is very weak and buggy.  Can it be trusted?
- Regressions to MingGW / GCC cross compilation.
- Did not explicitly follow Qt coding standard, but this can be easily
- addressed later.

================